requirements.txt
requests
prometheus_client
pyyaml
